refactor: move internal/utils package to pkg/utils
