+++ commands/ci/ciutils/utils.go
@@ -10,9 +10,9 @@ import (
 
 	"github.com/profclems/glab/pkg/iostreams"
 
-	"github.com/profclems/glab/internal/utils"
 	"github.com/profclems/glab/pkg/api"
 	"github.com/profclems/glab/pkg/tableprinter"
+	"github.com/profclems/glab/pkg/utils"
 
 	"github.com/pkg/errors"
 	"github.com/xanzy/go-gitlab"
