+++ commands/cmdutils/cmdutils.go
@@ -12,8 +12,8 @@ import (
 	"github.com/profclems/glab/pkg/iostreams"
 
 	"github.com/profclems/glab/internal/glrepo"
-	"github.com/profclems/glab/internal/utils"
 	"github.com/profclems/glab/pkg/api"
+	"github.com/profclems/glab/pkg/utils"
 	"github.com/xanzy/go-gitlab"
 
 	"github.com/AlecAivazis/survey/v2"
