+++ commands/issue/note/issue_note_create.go
@@ -7,8 +7,8 @@ import (
 	"github.com/profclems/glab/commands/issue/issueutils"
 
 	"github.com/profclems/glab/commands/cmdutils"
-	"github.com/profclems/glab/internal/utils"
 	"github.com/profclems/glab/pkg/api"
+	"github.com/profclems/glab/pkg/utils"
 
 	"github.com/spf13/cobra"
 	gitlab "github.com/xanzy/go-gitlab"
