+++ commands/issue/view/issue_view.go
@@ -9,8 +9,8 @@ import (
 	"github.com/profclems/glab/commands/issue/issueutils"
 
 	"github.com/profclems/glab/commands/cmdutils"
-	"github.com/profclems/glab/internal/utils"
 	"github.com/profclems/glab/pkg/api"
+	"github.com/profclems/glab/pkg/utils"
 
 	"github.com/MakeNowJust/heredoc"
 	"github.com/spf13/cobra"
