+++ commands/mr/note/mr_note_create.go
@@ -6,8 +6,8 @@ import (
 	"github.com/profclems/glab/commands/mr/mrutils"
 
 	"github.com/profclems/glab/commands/cmdutils"
-	"github.com/profclems/glab/internal/utils"
 	"github.com/profclems/glab/pkg/api"
+	"github.com/profclems/glab/pkg/utils"
 
 	"github.com/spf13/cobra"
 	"github.com/xanzy/go-gitlab"
