+++ commands/release/releaseutils/releaseutils.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/profclems/glab/pkg/iostreams"
 
-	"github.com/profclems/glab/internal/utils"
 	"github.com/profclems/glab/pkg/tableprinter"
+	"github.com/profclems/glab/pkg/utils"
 
 	"github.com/xanzy/go-gitlab"
 )
