+++ internal/glrepo/repo.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/profclems/glab/internal/git"
 	"github.com/profclems/glab/internal/glinstance"
-	"github.com/profclems/glab/internal/utils"
+	"github.com/profclems/glab/pkg/utils"
 	"github.com/xanzy/go-gitlab"
 )
